TST move sklearn vectorizer tests to their own test module
